--- src/human-readable/errors/index.ts
@@ -1,25 +1,0 @@
-export {
-  UnknownTypeError,
-  InvalidAbiItemError,
-  UnknownSolidityTypeError,
-} from './abiItem.js'
-
-export {
-  InvalidAbiTypeParameterError,
-  InvalidFunctionModifierError,
-  InvalidModifierError,
-  SolidityProtectedKeywordError,
-  InvalidParameterError,
-  InvalidAbiParametersError,
-  InvalidAbiParameterError,
-} from './abiParameter.js'
-
-export {
-  InvalidStructSignatureError,
-  InvalidSignatureError,
-  UnknownSignatureError,
-} from './signature.js'
-
-export { InvalidParenthesisError } from './splitParameters.js'
-
-export { CircularReferenceError } from './struct.js'
