+++ src/human-readable/parseAbiItem.ts
@@ -1,18 +1,13 @@
 import type { Abi } from '../abi.js'
+import { InvalidAbiItemError } from '../index.js'
 import type { Narrow } from '../narrow.js'
 import type { Error, Filter } from '../types.js'
-import { InvalidAbiItemError } from './errors/index.js'
-import {
-  isStructSignature,
-  parseSignature,
-  parseStructs,
-} from './runtime/index.js'
-import type {
-  ParseSignature,
-  ParseStructs,
-  Signature,
-  Signatures,
-} from './types/index.js'
+import { isStructSignature } from './runtime/signatures.js'
+import { parseStructs } from './runtime/structs.js'
+import { parseSignature } from './runtime/utils.js'
+import type { Signature, Signatures } from './types/signatures.js'
+import type { ParseStructs } from './types/structs.js'
+import type { ParseSignature } from './types/utils.js'
 
 /**
  * Parses human-readable ABI item (e.g. error, event, function) into {@link Abi} item
