--- src/human-readable/runtime/index.ts
@@ -1,10 +1,0 @@
-export {
-  eventModifiers,
-  functionModifiers,
-  isStructSignature,
-  modifiers,
-} from './signatures.js'
-
-export { parseStructs } from './structs.js'
-
-export { parseAbiParameter, parseSignature, splitParameters } from './utils.js'
