+++ src/human-readable/runtime/structs.ts
@@ -1,13 +1,13 @@
 import type { AbiParameter } from '../../abi.js'
 import { execTyped, isTupleRegex } from '../../regex.js'
-import { UnknownTypeError } from '../errors/index.js'
+import { UnknownTypeError } from '../errors/abiItem.js'
+import { InvalidAbiTypeParameterError } from '../errors/abiParameter.js'
 import {
-  CircularReferenceError,
-  InvalidAbiTypeParameterError,
   InvalidSignatureError,
   InvalidStructSignatureError,
-} from '../errors/index.js'
-import type { StructLookup } from '../types/index.js'
+} from '../errors/signature.js'
+import { CircularReferenceError } from '../errors/struct.js'
+import type { StructLookup } from '../types/structs.js'
 import { execStructSignature, isStructSignature } from './signatures.js'
 import { isSolidityType, parseAbiParameter } from './utils.js'
