--- src/human-readable/types/index.ts
@@ -1,18 +1,0 @@
-export type {
-  EventModifier,
-  FunctionModifier,
-  IsSignature,
-  IsStructSignature,
-  Modifier,
-  Signature,
-  Signatures,
-} from './signatures.js'
-
-export type { ParseStructs, StructLookup } from './structs.js'
-
-export type {
-  ParseAbiParameter,
-  ParseAbiParameters,
-  ParseSignature,
-  SplitParameters,
-} from './utils.js'
