+++ src/index.test.ts
@@ -11,20 +11,21 @@ it('should expose correct exports', () => {
       "parseAbiItem",
       "parseAbiParameter",
       "parseAbiParameters",
-      "CircularReferenceError",
-      "InvalidParenthesisError",
-      "UnknownSignatureError",
-      "InvalidSignatureError",
-      "InvalidStructSignatureError",
-      "InvalidAbiParameterError",
-      "InvalidAbiParametersError",
-      "InvalidParameterError",
-      "SolidityProtectedKeywordError",
-      "InvalidModifierError",
-      "InvalidFunctionModifierError",
-      "InvalidAbiTypeParameterError",
-      "InvalidAbiItemError",
       "UnknownTypeError",
+      "InvalidAbiItemError",
+      "UnknownSolidityTypeError",
+      "InvalidAbiTypeParameterError",
+      "InvalidFunctionModifierError",
+      "InvalidModifierError",
+      "SolidityProtectedKeywordError",
+      "InvalidParameterError",
+      "InvalidAbiParametersError",
+      "InvalidAbiParameterError",
+      "InvalidStructSignatureError",
+      "InvalidSignatureError",
+      "UnknownSignatureError",
+      "InvalidParenthesisError",
+      "CircularReferenceError",
     ]
   `)
 })
