+++ src/test.test.ts
@@ -1,6 +1,6 @@
 import { expect, it } from 'vitest'
 
-import * as Exports from './index.js'
+import * as Exports from './test.js'
 
 it('should expose correct exports', () => {
   expect(Object.keys(Exports)).toMatchInlineSnapshot(`
