+++ src/test.ts
@@ -12,7 +12,7 @@ export {
   wethAbi,
   writingEditionsFactoryAbi,
   eip165Abi,
-} from './abis.js'
+} from './test/abis.js'
 
 export {
   customSolidityErrorsHumanReadableAbi,
@@ -26,4 +26,4 @@ export {
   wethHumanReadableAbi,
   writingEditionsFactoryHumanReadableAbi,
   eip165HumanReadableAbi,
-} from './human-readable.js'
+} from './test/human-readable.js'
