--- src/zod/index.ts
@@ -1,23 +1,0 @@
-export {
-  Abi,
-  AbiConstructor,
-  AbiEvent,
-  AbiEventParameter,
-  AbiError,
-  AbiFallback,
-  AbiFunction,
-  AbiItemType,
-  AbiParameter,
-  AbiReceive,
-  AbiStateMutability,
-  SolidityAddress,
-  SolidityArray,
-  SolidityArrayWithoutTuple,
-  SolidityArrayWithTuple,
-  SolidityBool,
-  SolidityBytes,
-  SolidityFunction,
-  SolidityInt,
-  SolidityString,
-  SolidityTuple,
-} from './zod.js'
